+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New `!theme repo` will now send a message to chat with information about the extensions GitHub repo.
 - Added the `!theme help` command to explain to chat how to use the extensions commands
 
 ---